\name{getInterpolationMethodNames}
\alias{getInterpolationMethodNames}
\title{ get interpolation method names }
\description{ get interpolation method names }
\usage{ getInterpolationMethodNames() }
\details{ none }
\value{ 
character array with names for available interpolation methods
}


\references{ 

Pebesma, E., Cornford, D., Dubois, G., Heuvelink, G.B.M., Hristopulos, D., Pilz, J., Stohlker, U., Morin, G., Skoien, J.O. INTAMAP: The design and implementation f an interoperable automated interpolation Web Service. Computers and Geosciences 37 (3), 2011. 

}

\author{ Edzer Pebesma }

\examples{
getInterpolationMethodNames()
}
\keyword{spatial}
