\name{intamap}
\alias{observations}
\docType{data}
\title{Example data for the intamap package}
\description{
Salted (slightly modified) observations from the European gamma radiation network
}
\usage{data(intamap)}


\details{
The data set is a salted data set from the European gamma radiation network 
\url{https://remap.jrc.ec.europa.eu/GammaDoseRates.aspx}. 
Salted does in this context mean that all locations and observation values
have been randomized through addition of a random value.


}

\references{ 
Pebesma, E., Cornford, D., Dubois, G., Heuvelink, G.B.M., Hristopulos, D., Pilz, J., Stohlker, U., Morin, G., Skoien, J.O. INTAMAP: The design and implementation f an interoperable automated interpolation Web Service. Computers and Geosciences 37 (3), 2011. 
}

\keyword{datasets}
