\name{getInterpolationMethodNames}
\alias{getInterpolationMethodNames}
\title{ get interpolation method names }
\description{ get interpolation method names }
\usage{ getInterpolationMethodNames() }
\details{ none }
\value{ 
character array with names for available interpolation methods
}

\references{ 

\url{http://www.intamap.org/}

}
\author{ Edzer Pebesma }

\examples{
getInterpolationMethodNames()
}
\keyword{spatial}
