\name{intamap}
\alias{observations}
\docType{data}
\title{Example data for the intamap package}
\description{
Salted (slightly modified) observations from the European gamma radiation network
}
\usage{data(intamap)}


\details{
The data set is a salted data set from the European gamma radiation network 
\url{http://eurdep.jrc.ec.europa.eu/Basic/Pages/Public/Home/Default.aspx}. 
Salted does in this context mean that all locations and observation values
have been randomized through addition of a random value.


}


\keyword{datasets}
